requirements.txt
# Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Vector store
chromadb>=0.4.22

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.5

# Google Gemini
google-genai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web interface
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Development
pytest>=7.4.0
black>=24.0.0
ruff>=0.1.0
